Created a page to show posts for a specific topic
This page can be navigated to by the trending page. Clicking on the
links on the trending page takes you to this page which shows the posts
associated to the topiCreated a page to show posts for a specific topic

This page can be navigated to by the trending page. Clicking on the
links on the trending page takes you to this page which shows the posts
associated to the topic